Adds implementation of internal HG structures
This commit adds the implementation of each function of the different structures defined in `HGInternalStruct.h` in the precedent commits. Thus all needed functionalities which handle the internal representation of a room as a grid are now working (it is easier to create the source to generate and spawn the furniture and the doors). Furthermore, some structures have been moved in `HGBasicType.h` (resp. `HGBasicType.cpp`) to avoid circular inclusion (actually just `FVectorGrid`).