Adding room division steps
This commits first adds needed internal structs to store the rooms beeing divided/splited during the BSP.
The HomeGenerator class divides the surface of each level after choosing the place of the first halls and the place of the first blocks (same for all levels). It then divides the surface of each level and finally allocate on all building's rooms their type.
Actually no wall are generated only the abstracts datas.